backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic
pydantic-settings
python-dotenv
python-multipart

sqlalchemy
alembic
psycopg2-binary

openai
httpx
pypdf
chromadb
pgvector
supabase
slowapi